% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiscPlot.R
\name{DiscDots}
\alias{DiscDots}
\title{Create discrete dot plot}
\usage{
DiscDots(values, var_names, ...)
}
\arguments{
\item{values}{Number of  filled in dots for each variable}
}
\value{
dot plot
}
\description{
Create discrete dot plot
}

