% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiscPlot.R
\name{gen_df}
\alias{gen_df}
\title{Create Gauge Dataframe}
\usage{
gen_df(values, ...)
}
\arguments{
\item{values}{Number of  filled in dots for each variable}
}
\value{
dataframe of 2's and 1's to be fed into plotting functions.
}
\description{
Create Gauge Dataframe
}

