% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiscPlot.R
\name{gen_mat}
\alias{gen_mat}
\title{Create Gauge Matrix}
\usage{
gen_mat(values, var_names, ...)
}
\arguments{
\item{values}{Number of  filled in dots for each variable}
}
\value{
matrix of 2's and 1's to be fed into plotting functions.
}
\description{
Create Gauge Matrix
}

