% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DiscPlot.R
\name{theme_ddots}
\alias{theme_ddots}
\title{Default theme for DiscDots}
\usage{
theme_ddots(base_size = 26, base_family = "Josefin Sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme generates the style of plot in the documentation/github example
}

